+++ include/shape.hpp
@@ -264,8 +264,8 @@ public:
 		unsigned int maxMipLevels = 5;
 
 		for (unsigned int mip = 0; mip < maxMipLevels; ++mip){
-			unsigned int mipWidth  = 128 * std::pow(0.5, mip);
-			unsigned int mipHeight = 128 * std::pow(0.5, mip);
+			unsigned int mipWidth  = static_cast<unsigned int>(128 * std::pow(0.5, mip));
+			unsigned int mipHeight = static_cast<unsigned int>(128 * std::pow(0.5, mip));
 			glBindRenderbuffer(GL_RENDERBUFFER, captureRBO);
 			glRenderbufferStorage(GL_RENDERBUFFER, GL_DEPTH_COMPONENT24, mipWidth, mipHeight);
 			glViewport(0, 0, mipWidth, mipHeight);
